Support plus unary operator
